chore(plugins/plugin-client-common): extract sourceRef out of Input.tsx
Fixes #6860
